// dist/Attribute/AttributeType.d.ts
export declare enum AttributeType {
    /**
     * Continuous Attribute
     */
    CONTINUOUS = 0,
    /**
     * Discrete Attribute
     */
    DISCRETE = 1,
    /**
     * Binary Attribute
     */
    BINARY = 2,
    /**
     * Discrete Indexed Attribute is used to store the indices.
     */
    DISCRETE_INDEXED = 3
}